refactor: remove all mentions of eralchemy2
