+++ .github/workflows/python-app.yml
@@ -72,4 +72,4 @@ jobs:
         run: ruff format --check
 
       - name: Run mypy
-        run: mypy eralchemy2
+        run: mypy eralchemy
