+++ example/graph.py
@@ -4,7 +4,7 @@
 from sqlalchemy import Column, ForeignKey, Integer
 from sqlalchemy.orm import declarative_base, relationship
 
-from eralchemy2 import render_er
+from eralchemy import render_er
 
 Base = declarative_base()
